pyproject.toml
[project]
name = "qslweb"
version = "0.1.0"
description = "A Django application for managing eQSL card confirmations via email"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1,<6.0",
    "django-q2>=1.7.0",
    "pillow>=10.0.0",
    "psycopg2-binary>=2.9.0",
    "redis>=5.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.5.0",
    "playwright>=1.40.0",
    "pytest-playwright>=0.4.0",
    "ruff>=0.6.0",
    "ipython>=8.20.0",
    "django-debug-toolbar>=4.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["qslweb"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "qslweb.settings"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--showlocals",
    "--tb=short",
]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "ui: marks tests as UI tests using playwright",
]

[tool.coverage.run]
source = ["."]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/venv/*",
    "*/virtualenv/*",
    ".venv/*",
    "manage.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "DJ",     # flake8-django
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "DJ001",  # avoid using null=True on string-based fields
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"*/migrations/*" = ["ALL"]
"*/tests/*" = ["ARG", "S"]

[tool.ruff.lint.isort]
known-first-party = ["qslweb"]

[tool.django-stubs]
django_settings_module = "qslweb.settings"
